Better code sharing with similar classes
